+++ src/bricks/transformers/temporaryInfo/DisplayTemporaryInfo.ts
@@ -27,6 +27,7 @@ import {
   hideTemporarySidebarPanel,
   showTemporarySidebarPanel,
   updateTemporarySidebarPanel,
+  onSidePanelClosure,
 } from "@/contentScript/sidebarController";
 import {
   type PanelPayload,
@@ -55,7 +56,6 @@ import { TransformerABC } from "@/types/bricks/transformerTypes";
 import { type Schema } from "@/types/schemaTypes";
 import { type Location } from "@/types/starterBrickTypes";
 import { assumeNotNullish_UNSAFE } from "@/utils/nullishUtils";
-import { onSidePanelClosure } from "@/sidebar/sidePanel/messenger/api";
 
 // Match naming of the sidebar panel extension point triggers
 export type RefreshTrigger = "manual" | "statechange";
